move ArraySerializer tests into separate file
